# improve adler32 perf