casts length variable into i32

|| 1gb | 500mb |
|--|--| --|
| before | 1134ms | 553ms |
| after | 845ms | 412ms |