@101arrowz 2021-04-27
Looks good. I'll probably test this a bit further before merging though - I think this optimization can be applied elsewhere too.